My prorfiloi/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
Jinja2==3.1.2
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0
Flask-Mail==0.9.1